Move lightning tests to notravis.
