+++ tests/unit/algorithms/lightning_anil_test_notravis.py
@@ -1,21 +1,31 @@
 #!/usr/bin/env python3
 
 import unittest
+import torch
 import learn2learn as l2l
 import pytorch_lightning as pl
 
 from learn2learn.utils.lightning import EpisodicBatcher
-from learn2learn.algorithms import LightningPrototypicalNetworks
+from learn2learn.algorithms import LightningANIL
 
 
-class TestLightningProtoNets(unittest.TestCase):
+class Lambda(torch.nn.Module):
+    def __init__(self, fn):
+        super(Lambda, self).__init__()
+        self.fn = fn
 
-    def test_protonets(self):
-        meta_batch_size = 4
-        max_epochs = 20
+    def forward(self, x):
+        return self.fn(x)
+
+
+class TestLightningANIL(unittest.TestCase):
+    def test_anil(self):
+        meta_batch_size = 32
+        max_epochs = 5
         seed = 42
         ways = 5
         shots = 5
+        adaptation_lr = 5e-1
 
         pl.seed_everything(seed)
 
@@ -29,10 +39,13 @@ class TestLightningProtoNets(unittest.TestCase):
             root="~/data",
         )
 
+        self.assertTrue(len(tasksets) == 3)
+
+        features = l2l.vision.models.ConvBase(output_size=64, channels=3, max_pool=True)
+        features = torch.nn.Sequential(features, Lambda(lambda x: x.view(-1, 256)))
+        classifier = torch.nn.Linear(256, ways)
         # init model
-        model = l2l.vision.models.CNN4(ways, embedding_size=32*4)
-        features = model.features
-        protonet = LightningPrototypicalNetworks(features, lr=3e-4)
+        anil = LightningANIL(features, classifier, adaptation_lr=adaptation_lr)
         episodic_data = EpisodicBatcher(
             tasksets.train,
             tasksets.validation,
@@ -42,12 +55,13 @@ class TestLightningProtoNets(unittest.TestCase):
 
         trainer = pl.Trainer(
             accumulate_grad_batches=meta_batch_size,
+            min_epochs=max_epochs,
             max_epochs=max_epochs,
             progress_bar_refresh_rate=0,
             deterministic=True,
             weights_summary=None,
         )
-        trainer.fit(protonet, episodic_data)
+        trainer.fit(anil, episodic_data)
         acc = trainer.test(
             test_dataloaders=tasksets.test,
             verbose=False,
